requirements.txt
Django==1.6.7
South==1.0.1
wsgiref==0.1.2
django-rq==0.7.0
Pillow==2.6.1
django-cities-light==2.3.0
django-oauth2-provider==0.2.6.1
djangorestframework==2.4.4
beautifulsoup4==4.3.2
feedparser==5.1.3
mysolr==0.8.3
readability-lxml==0.3.0.5
python-dateutil==2.2
django-rest-swagger==0.2.1
django-bootstrap3==4.11.0
sorl-thumbnail==12.1c
pytz>=2014.9
lxml==3.4.1
requests==2.4.3
django-cache-machine==0.8
django-rq-dashboard==0.2.2
urlnorm==1.1.2
django-markdown-deux==1.0.5
html2text==3.200.3
#psycopg2
#pylibmc